move multispinner into its own file
